+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Update Distrowatch parser.
 - Fix plugins sometimes not getting loaded.
+- Fix config command.
 
 ## [1.1.0] - 2018-05-16
 ### Added